backup_requirement.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.4
ldap3==2.9.1
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
pyasn1==0.4.8
pyotp==2.8.0
python-dateutil==2.8.2
pytz==2022.2.1
qrcode==7.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12




django-filter==22.1
Django==4.1.1
djangorestframework==3.13.1